refactor: 컨벤션 및 패키지 구조 변경
