+++ src/main/java/servlet/domain/user/UserCreateServlet.java
@@ -1,4 +1,4 @@
-package servlet;
+package servlet.domain.user;
 
 import java.util.Map;
