+++ src/main/java/webserver/WebServer.java
@@ -12,36 +12,37 @@ import container.MyContainer;
 
 public class WebServer {
 
-    private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
-    private static final int DEFAULT_PORT = 8080;
-    private static final int MAX_THREADS = 100;
+	private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
+	private static final int DEFAULT_PORT = 8080;
+	private static final int MAX_THREADS = 100;
 
-    public static void main(String args[]) throws Exception {
-        int port = 0;
-        if (args == null || args.length == 0) {
-            port = DEFAULT_PORT;
-        } else {
-            port = Integer.parseInt(args[0]);
-        }
+	public static void main(String args[]) throws Exception {
+		int port = 0;
+		if (args == null || args.length == 0) {
+			port = DEFAULT_PORT;
+		} else {
+			port = Integer.parseInt(args[0]);
+		}
 
-        MyContainer.start(WebServer.class);
+		MyContainer.start(WebServer.class);
 
-        ExecutorService executor = Executors.newFixedThreadPool(MAX_THREADS);
+		ExecutorService executor = Executors.newFixedThreadPool(MAX_THREADS);
 
-        // 서버소켓을 생성한다. 웹서버는 기본적으로 8080번 포트를 사용한다.
-        try (ServerSocket listenSocket = new ServerSocket(port)) {
-            logger.info("Web Application Server started {} port.", port);
+		// 서버소켓을 생성한다. 웹서버는 기본적으로 8080번 포트를 사용한다.
+		try (ServerSocket listenSocket = new ServerSocket(port)) {
+			logger.info("Web Application Server started {} port.", port);
 
-            // 클라이언트가 연결될때까지 대기한다.
-            Socket connection;
-            while ((connection = listenSocket.accept()) != null) {
-                Runnable worker = new RequestHandler(connection);
-                executor.execute(worker);
-            }
-        }
+			// 클라이언트가 연결될때까지 대기한다.
+			Socket connection;
+			while ((connection = listenSocket.accept()) != null) {
+				Runnable worker = new RequestHandler(connection);
+				executor.execute(worker);
+			}
+		}
 
-        executor.shutdown();
-        while (!executor.isTerminated()) { }
-        logger.info("Server closed...");
-    }
+		executor.shutdown();
+		while (!executor.isTerminated()) {
+		}
+		logger.info("Server closed...");
+	}
 }
