+++ src/main/java/webserver/utils/view/FileUtil.java
@@ -5,21 +5,23 @@ import java.util.Optional;
 
 public class FileUtil {
 
-    public static boolean isFileRequest(String url) {
-        String[] splitUrl = url.split("[.]");
-        if(splitUrl.length == 0) return false;
+	public static boolean isFileRequest(String url) {
+		String[] splitUrl = url.split("[.]");
+		if (splitUrl.length == 0) {
+			return false;
+		}
 
-        String fileExtension = splitUrl[splitUrl.length - 1];
+		String fileExtension = splitUrl[splitUrl.length - 1];
 
-        return isProvidedExtension(fileExtension);
-    }
+		return isProvidedExtension(fileExtension);
+	}
 
-    private static boolean isProvidedExtension(String extension) {
-        Optional<String> findExtension = Arrays.stream(Extension.values())
-                .map(t -> t.getValue())
-                .filter(s -> s.equals(extension))
-                .findFirst();
+	private static boolean isProvidedExtension(String extension) {
+		Optional<String> findExtension = Arrays.stream(Extension.values())
+			.map(t -> t.getValue())
+			.filter(s -> s.equals(extension))
+			.findFirst();
 
-        return findExtension.isPresent();
-    }
+		return findExtension.isPresent();
+	}
 }
